+++ src/main/java/com/billit/user_service/account/domain/repository/InvestAccountRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 
 public interface InvestAccountRepository extends JpaRepository<InvestAccount, Long> {
     List<InvestAccount> findAllByUserInvestIdAndIsDeletedFalse(Long userInvestId);
-    Optional<InvestAccount> findByIdAndIsDeletedFalse(Long id);
+    Optional<InvestAccount> findByIdAndIsDeletedFalse(Integer id);
     Optional<InvestAccount> findByAccountNumberAndIsDeletedFalse(String accountNumber);
     boolean existsByAccountNumberAndIsDeletedFalse(String accountNumber);
 }
